+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - Introduced [sentry](https://www.sentry.io) to capture exception and upload them to sentry server automatically, which
   helps us to track and fix bug more efficiently. Sentry is only enabled in non-dev version.
 - Server version will also be uploaded to bStats now.
+- Introduced `Extension#afterServerStarted` method which will be called after the server is started.
 
 ### Changed